requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
pydantic
python-dotenv